Updated profile.ps1 - New-RebootScheduledTask.ps1

+++ profile.ps1
@@ -1572,7 +1572,7 @@ Function New-RebootScheduledTask {
         [Parameter()]
         [String]$ComputerName,
         [Parameter(Mandatory)]
-        [Datetime]$Time,
+        [Datetime]$UTCTime,
         [Parameter(Mandatory)]
         [String]$Description,
         [Parameter()]
@@ -1591,6 +1591,10 @@ Function New-RebootScheduledTask {
         ErrorAction = "SilentlyContinue"
     }
 
+    $GetCimInstanceSplat = @{
+        ErrorAction = "Stop"
+    }
+
     if ($PSBoundParameters.ContainsKey("ComputerName")) {
         $NewCimSession = @{
             ComputerName = $ComputerName
@@ -1600,6 +1604,7 @@ Function New-RebootScheduledTask {
             $NewCimSession["Credential"] = $Credential
         }
         $Session = New-CimSession @NewCimSession
+        $GetCimInstanceSplat["CimSession"] = $Session
         $GetScheduledTaskSplat["CimSession"] = $Session
     }
 
@@ -1622,13 +1627,40 @@ Function New-RebootScheduledTask {
         }
     }
 
-    $Description = "{0} - created by {1} on {2}" -f $Description, $env:USERNAME, (Get-Date)
+    try {
+        $TimeZone = Get-CimInstance -ClassName "Win32_TimeZone" @GetCimInstanceSplat
+    }
+    catch {
+        Write-Warning "Could not determine target system's timezone"
+    }
+
+    try {
+        $LocalTime = Get-CimInstance -ClassName "Win32_LocalTime" @GetCimInstanceSplat
+        $LocalTime = Get-Date -Year $LocalTime.year -Month $LocalTime.month -day $LocalTime.day -Hour $LocalTime.hour -Minute $LocalTime.minute -Second $LocalTime.Second
+    }
+    catch {
+        Write-Warning "Could not detemrine target system's local time"
+    }
+
+    $Y = New-Object System.Management.Automation.Host.ChoiceDescription "&Yes", "Continue with lab build"
+    $N = New-Object System.Management.Automation.Host.ChoiceDescription "&No", "Do not continue with lab build"
+    $Options = [System.Management.Automation.Host.ChoiceDescription[]]($Y, $N)
+    $Message = "Target machine's timezone is '{0}' ('{1}'), would you like to continue with '{2} UTC?'" -f $TimeZone.Caption, $LocalTime, $UTCTime.ToUniversalTime()
+    $Result = $host.ui.PromptForChoice($null, $Message, $Options, 1)
+
+    if ($Result -ge 1) {
+        return
+    }
+
+    $Description = "{0} - created by {1} on {2} (UTC)" -f $Description, $env:USERNAME, $UTCTime.ToUniversalTime()
 
     $Action = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument '-NonInteractive -NoLogo -NoProfile -Command "Restart-Computer -Force"'
-    $Trigger = New-ScheduledTaskTrigger -Once -At $Time
+    $Trigger = New-ScheduledTaskTrigger -Once -At $UTCTime.ToUniversalTime()
     $Settings = New-ScheduledTaskSettingsSet
     $Task = New-ScheduledTask -Action $Action -Trigger $Trigger -Settings $Settings -Description $Description
     Register-ScheduledTask -TaskName $TaskName -TaskPath $TaskPath -InputObject $Task -User "System" -CimSession $Session
+
+    if ($Session) { Remove-CimSession -CimSession $Session }
 }
 
 Function Get-Dns {
